Move `enum_map` tests to integration tests
This also ensures that they only run when the `macros` feature is provided.
